bruh

+++ app/src/main/java/com/example/androidtrainingray/ResultWeather.kt
@@ -23,7 +23,7 @@ data class ResultWeather(
 data class ListItem(
 
 	@field:SerializedName("dt")
-	val dt: Int? = null,
+	val dt: Long? = null,
 
 	@field:SerializedName("pop")
 	val pop: Double? = null,
